// src/types/pic.d.ts
export interface Project {
  /**
   * database id
   */
  id: number;
  /**
   * Federal agency that supervises preparation of the environmental documents for the project.
   */
  lead_agency?: string;
  /**
   * Other involved agencies.
   */
  participating_agencies?: string;
  /**
   * Geographic area including GIS data
   */
  location_object?: {
    /**
     * Text description of the project location
     */
    description: string;
    /**
     * Reference to associated GIS data
     */
    gis_reference?: string;
    [k: string]: unknown;
  };
  /**
   * Environmental review initiation date (aggregated from process status)
   */
  start_date?: string;
  /**
   * Current phase (pre-application, underway, paused, completed) of the NEPA, permitting, or other authorization process (aggregated from process status).
   */
  current_status?: "pre-application" | "underway" | "paused" | "completed";
  /**
   * Center/centroid of project (shortcut for plotting on a map - should attach GIS data object as well)
   */
  location_lat?: number;
  /**
   * Center/centroid of project (shortcut for plotting on a map - should attach GIS data object as well)
   */
  location_lon?: number;
  /**
   * Name of responsible entity, organization, or person (contact information json object)
   */
  sponsor_contact?: {
    [k: string]: unknown;
  };
  /**
   * This is the version of this data record.
   */
  data_record_version?: string;
  /**
   * This is the agency from which this data record was sent or retrieved or in which it is stored, which may be different than the record owner.
   */
  data_source_agency?: string;
  /**
   * This is the system from which this data record was sent or retrieved or in which it is stored.
   */
  data_source_system?: string;
  /**
   * This is the last update of the data record in the system.
   */
  last_updated?: string;
  /**
   * This is the owner of the original data, generally the authoritative data source.
   */
  record_owner_agency?: string;
  /**
   * This is the timestamp of a retrieval or transmission (e.g. responding to an API call).
   */
  retrieved_timestamp?: string;
  /**
   * created
   */
  created_at?: string;
  /**
   * Descriptive name of the project.
   */
  title: string;
  /**
   * Summary of the project’s goals and scope.
   */
  description: string;
  /**
   * High-level project category (e.g., energy, transportation, land management).
   */
  sector?: string;
  /**
   * Classification, a sub-type of sector (pipeline, highway, habitat restoration).
   */
  type?: string;
  /**
   * Link to federal, state, local or other funding by reference (e.g., grant program, loan guarantee, appropriated funds)
   */
  funding?: string;
  /**
   * Name of responsible entity, organization, or person.
   */
  sponsor: string;
  /**
   * If nesting projects, reference to the parent project's ID.
   */
  parent_project_id?: number;
  /**
   * Text field to specify location, eg. address or Well Known Text, WKT
   */
  location_text?: string;
  /**
   * This jsonb field should be used for additional data structures as needed.
   */
  other?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
/**
 * The specific type of environmental review, permit, or authorization. cA process is associated with or nested beneath a project and will have documents associated and nested beneath it.
 */
export interface ProcessInstance {
  /**
   * Database primary key identifier
   */
  id: number;
  /**
   * created
   */
  created_at?: string;
  /**
   * Reference to Process ID of parent process, if applicable
   */
  parent_process_id?: number;
  /**
   * Reference to Project ID of parent process
   */
  parent_project_id: number;
  /**
   * Process-specific ID assigned by agency, if assigned
   */
  agency_id?: string;
  /**
   * Unique identification number assigned at federal level
   */
  federal_id?: string;
  /**
   * Level of NEPA review or other permit or authorization (see Permitting Council Federal Environmental Review and Authorization Inventory)
   */
  type: string;
  /**
   * Current status of the process
   */
  status: ("planned" | "underway" | "paused" | "completed") | string;
  /**
   * Current state of process, may be derived from the last case event with this Process ID
   */
  stage?: string;
  /**
   * Initiation date of the process
   */
  start_date?: string;
  /**
   * Conclusion date of the process
   */
  complete_date?: string;
  /**
   * Result (e.g., Record of Decision, permit issuance)
   */
  outcome?: string;
  /**
   * Federal agency that supervises preparation of the environmental documents for the project
   */
  lead_agency: string;
  /**
   * Joint lead agency, if applicable
   */
  joint_lead_agency?: string;
  /**
   * Cooperating agencies involved in the process
   */
  cooperating_agencies?: string[];
  /**
   * Other agencies involved in the process
   */
  participating_agencies?: string[];
  /**
   * Start date and time of public comment period
   */
  comment_start?: string;
  /**
   * End date and time of public comment period
   */
  comment_end?: string;
  /**
   * Purpose and need statement for the project
   */
  purpose_need?: string;
  /**
   * Miscellaneous notes on process instance model
   */
  notes?: string;
  /**
   * This jsonb field should be used for additional data structures as needed.
   */
  other?: {
    [k: string]: unknown;
  };
  /**
   * Detailed description of the process
   */
  description?: string;
  /**
   * Reference to the process model used for this process instance
   */
  process_model?: string | null;
  /**
   * This is the owner of the original data, generally the authoritative data source.
   */
  record_owner_agency?: string;
  /**
   * This is the agency from which this data record was sent or retrieved or in which it is stored, which may be different than the record owner.
   */
  data_source_agency?: string;
  /**
   * This is the system from which this data record was sent or retrieved or in which it is stored.
   */
  data_source_system?: string;
  /**
   * This is the version of this data record.
   */
  data_record_version?: string;
  /**
   * This is the last update of the data record in the system.
   */
  last_updated?: string;
  /**
   * This is the timestamp of a retrieval or transmission (e.g. responding to an API call).
   */
  retrieved_timestamp?: string;
  /**
   * Short internal process code used for mapping process types to external or legacy systems.
   */
  process_code?: string;
  [k: string]: unknown;
}
/**
 * Official records created during the NEPA review. The document object includes both metadata identifying the document and its context and a container for summary information that lays out, at a minimum, the table of contents of the document. Documents are also likely to have many GIS objects associated with them.
 */
export interface Document {
  /**
   * database id
   */
  id: number;
  /**
   * Document category, such as EIS, EA, etc.
   */
  document_type: string;
  /**
   * Document title.
   */
  title: string;
  /**
   * Volume title (e.g., Appendix)
   */
  volume_title?: string;
  /**
   * Publication or submission date.
   */
  publish_date?: string;
  /**
   * Responsible entity.
   */
  prepared_by: string;
  /**
   * Document production phase.
   */
  status?: string;
  /**
   * Online link if available.
   */
  url?: string;
  /**
   * Miscellaneous notes.
   */
  notes?: string;
  /**
   * Unique identification number for related document (if applicable)
   */
  related_document_id?: number;
  /**
   * container for summary information
   */
  document_summary?: {
    [k: string]: unknown;
  };
  /**
   * the table of contents (or heading structure) of the document.
   */
  document_toc?: {
    [k: string]: unknown;
  };
  /**
   * Public or restricted.
   */
  public_access?: boolean;
  /**
   * This is the version of this data record.
   */
  data_record_version?: string;
  /**
   * This is the agency from which this data record was sent or retrieved or in which it is stored, which may be different than the record owner.
   */
  data_source_agency?: string;
  /**
   * This is the system from which this data record was sent or retrieved or in which it is stored.
   */
  data_source_system?: string;
  /**
   * This is the last update of the data record in the system.
   */
  last_updated?: string;
  /**
   * This is the owner of the original data, generally the authoritative data source.
   */
  record_owner_agency?: string;
  /**
   * This is the timestamp of a retrieval or transmission (e.g. responding to an API call).
   */
  retrieved_timestamp?: string;
  /**
   * This jsonb field should contain any document files if applicable; this is an optional way of storing or transmitting files and is not necessary as this is primarily a data record for metadata about the document rather than the document itself.
   */
  document_files?: {
    [k: string]: unknown;
  };
  /**
   * creation date
   */
  created_at?: string;
  /**
   * Reference to Process ID of parent process for the document
   */
  parent_process_id: number;
  /**
   * Indicates which revision of the document (e.g., first revised Draft EIS) - numeric form
   */
  revision_number?: number;
  /**
   * Indicates which revision of the document (e.g., first revised Draft EIS) - numeric form
   */
  document_revision?: number;
  /**
   * Indicates supplement number
   */
  supplement_number?: number;
  /**
   * This jsonb field should be used for additional data structures as needed.
   */
  other?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
// ...
export interface CaseEvent {
  /**
   * database id
   */
  id: number;
  /**
   * Link to information about the event.
   */
  source?: string;
  /**
   * Event class (e.g., NOI, ROD)
   */
  type: string;
  /**
   * Optional event hierarchy (e.g., Tier 1 NOI, Tier 2 Scoping Complete, Tier 3 Task assigned to Jane Doe).
   */
  tier?: number;
  /**
   * Pending, completed, in progress.
   */
  status?: "pending" | "completed" | "in progress";
  /**
   * Result or action taken.
   */
  outcome?: string;
  /**
   *  Reference to Document ID of the document related to the event. Optional.
   */
  related_document_id?: number;
  /**
   * Indicates whether events can be displayed publicly. (Binary Y/N)
   */
  public_access?: boolean;
  /**
   * Description of event.
   */
  description?: string;
  /**
   * Responsible individuals or agencies.
   */
  assigned_entity?: string;
  /**
   * Name of segment following this event (e.g. if this is "start of scoping" the next segment is "scoping").
   */
  following_segment_name?: string;
  /**
   * If applicable, the parent ID of the event.
   */
  parent_event_id?: number;
  /**
   * Related engagement event if applicable
   */
  related_engagement_id?: number;
  /**
   * This is the version of this data record.
   */
  data_record_version?: string;
  /**
   * This is the agency from which this data record was sent or retrieved or in which it is stored, which may be different than the record owner.
   */
  data_source_agency?: string;
  /**
   * This is the system from which this data record was sent or retrieved or in which it is stored.
   */
  data_source_system?: string;
  /**
   * This is the last update of the data record in the system.
   */
  last_updated?: string;
  /**
   * This is the owner of the original data, generally the authoritative data source.
   */
  record_owner_agency?: string;
  /**
   * This is the timestamp of a retrieval or transmission (e.g. responding to an API call).
   */
  retrieved_timestamp?: string;
  /**
   * created
   */
  created_at?: string;
  /**
   * Reference to Process ID of the associated NEPA process for this event.
   */
  parent_process_id: number;
  /**
   * Name of event
   */
  name: string;
  /**
   * Date and time of event
   */
  datetime: string;
  /**
   * This jsonb field should be used for additional data structures as needed.
   */
  other?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}